thin out landingzone.go a little
